Moved subclassed widgets to ui folder, subclassed listwidget
