+++ ui/__init__.py
@@ -1,0 +1,3 @@
+from ui.UiProfileItem import UiProfileItem
+from ui.UiListWidget import UiListWidget
+from ui.MainWindow import Ui_MainWindow
